LayoutTests/fast/mediastream/mediastream-onaddtrack-and-overconstrainederror-code.html

<!doctype html>
<html>
<head>
<meta charset="utf-8">
<title>MediaStream.onaddtrack IDL handler + OverconstrainedError.code is 0</title>
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<script>
// MediaCapture-Main §"MediaStream" and §"OverconstrainedError":
//   https://w3c.github.io/mediacapture-main/#dom-mediastream-onaddtrack
//   https://w3c.github.io/mediacapture-main/#dom-mediastream-onremovetrack
//   https://w3c.github.io/mediacapture-main/#overconstrainederror-interface
//
// Two unrelated IDL-shape behaviors tested together because each is a
// one-liner:
//
//   MediaStream.onaddtrack — event handler IDL attribute for addtrack events.
//   OverconstrainedError.code is 0 (the interface is NOT in the legacy
//   DOMException name -> code mapping).

if (window.testRunner) {
    testRunner.setMicrophonePermission(true);
    testRunner.setCameraPermission(true);
}

test(() => {
    // MediaStream.onaddtrack IDL attribute exists, defaults to null,
    // and is settable to a function.
    const ms = new MediaStream();
    assert_true("onaddtrack" in ms,
        "MediaStream exposes onaddtrack IDL attribute");
    assert_equals(ms.onaddtrack, null,
        "onaddtrack defaults to null");
    const handler = () => {};
    ms.onaddtrack = handler;
    assert_equals(ms.onaddtrack, handler,
        "onaddtrack setter accepts a function and getter returns it");

    // Same shape for onremovetrack (already covered by an existing
    // WebRTC test for the firing path; here we verify the IDL surface).
    assert_true("onremovetrack" in ms,
        "MediaStream exposes onremovetrack IDL attribute");
    assert_equals(ms.onremovetrack, null,
        "onremovetrack defaults to null");
}, "MediaStream.onaddtrack and onremovetrack IDL attributes (existence + default)");

promise_test(async test => {
    // Firing path: addtrack fires when the UA adds a track to a
    // stream. We construct a fresh empty MediaStream, register onaddtrack,
    // then add a track via the script-side addTrack API. Per spec, the
    // UA is supposed to fire addtrack only for UA-initiated additions
    // (e.g. PeerConnection remote tracks); script-initiated addTrack
    // historically does fire addtrack in WebKit. Test what WebKit does:
    // either way, the event payload must be a TrackEvent whose .track
    // matches the added track.
    const source = await navigator.mediaDevices.getUserMedia({ audio: true });
    test.add_cleanup(() => source.getTracks().forEach(t => t.stop()));
    const audioTrack = source.getAudioTracks()[0];

    const ms = new MediaStream();
    let firedEvent = null;
    ms.onaddtrack = e => { firedEvent = e; };

    ms.addTrack(audioTrack);

    // Give any UA-queued addtrack task time to fire.
    await new Promise(r => test.step_timeout(r, 100));

    if (firedEvent === null) {
        // WebKit follows the strict spec reading: addtrack only fires
        // for UA-initiated additions. The IDL handler exists and is
        // settable (subtest above), but doesn't fire for script
        // addTrack. This is spec-conforming.
        assert_true(ms.getTracks().includes(audioTrack),
            "track was added even though addtrack did not fire");
    } else {
        // WebKit fires addtrack for script-initiated addTrack — also
        // observed in practice across browsers. Verify the event
        // shape.
        assert_true(firedEvent instanceof Event,
            "addtrack event is an Event instance");
        assert_equals(firedEvent.type, "addtrack",
            "event type is 'addtrack'");
        assert_equals(firedEvent.track, audioTrack,
            "TrackEvent.track is the added track");
    }
}, "MediaStream.onaddtrack: addtrack event payload (when fired by UA or by script addTrack)");

test(() => {
    // OverconstrainedError.code is 0. OverconstrainedError extends
    // DOMException. DOMException.code is the legacy numeric code mapped
    // from .name via a fixed table — names not in the table get code 0.
    // OverconstrainedError's .name is "OverconstrainedError", which is
    // NOT in the legacy mapping, so .code must be 0.
    const err = new OverconstrainedError("width", "no camera satisfies the width constraint");
    assert_true(err instanceof DOMException,
        "OverconstrainedError inherits from DOMException");
    assert_equals(err.name, "OverconstrainedError",
        "name is 'OverconstrainedError'");
    assert_equals(err.code, 0,
        "code is 0 (no legacy DOMException code mapping for OverconstrainedError)");
    assert_equals(err.constraint, "width",
        "constraint attribute is set from the constructor");
    assert_equals(err.message, "no camera satisfies the width constraint",
        "message attribute is set from the constructor");
}, "OverconstrainedError.code is 0 (no legacy DOMException code mapping)");

promise_test(async test => {
    // Real path: an OverconstrainedError raised by the UA from
    // applyConstraints must also have .code === 0.
    const stream = await navigator.mediaDevices.getUserMedia({ video: true });
    test.add_cleanup(() => stream.getTracks().forEach(t => t.stop()));
    const track = stream.getVideoTracks()[0];

    let raisedError = null;
    try {
        // exact: impossibly-large width — no real camera satisfies.
        await track.applyConstraints({ width: { exact: 99999999 } });
        assert_unreached("applyConstraints with impossible exact width should reject");
    } catch (e) {
        raisedError = e;
    }
    assert_equals(raisedError.name, "OverconstrainedError",
        "rejection is an OverconstrainedError");
    assert_equals(raisedError.code, 0,
        "UA-raised OverconstrainedError.code is also 0");
}, "OverconstrainedError raised by applyConstraints also has code 0");
</script>
</head>
<body></body>
</html>
